+++ vorx/conf.py
@@ -1,4 +1,4 @@
-VERSION = "2026.0.5.1"
+VERSION = "2026.0.5.2"
 BUILD = "dev"
-BUILD_DATE = "2026-05-17"
-SERVER_SOCK = "vorx.codcatdev.site"
+BUILD_DATE = "2026-05-22"
+PING_SOCK = "https://api.vorx.pro/v1/status"